+++ src/editors/object.js
@@ -637,7 +637,7 @@ export class ObjectEditor extends AbstractEditor {
       this.addproperty_holder.appendChild(spacer)
 
       /* Close properties modal if clicked outside modal */
-      document.addEventListener('click', this.onOutsideModalClick)
+      document.addEventListener('click', this.onOutsideModalClick.bind(this))
 
       /* Description */
       if (this.schema.description) {
@@ -1042,9 +1042,8 @@ export class ObjectEditor extends AbstractEditor {
   }
 
   onOutsideModalClick (e) {
-    if (this.addproperty_holder &&
-      !this.addproperty_holder.contains(e.path[0] || e.composedPath()[0]) &&
-      this.adding_property) {
+    const path = e.path || (e.composedPath && e.composedPath())
+    if (this.addproperty_holder && !this.addproperty_holder.contains(path[0]) && this.adding_property) {
       e.preventDefault()
       e.stopPropagation()
       this.toggleAddProperty()